Use ActiveRecord::Concern
