+++ app/decorators/spree/products_controller_decorator.rb
@@ -1,0 +1,1 @@
+Spree::ProductsController.prepend(SolidusProductFeed::LoadFeedProductsConcern)
